refactored priors into package
