+++ autofit/mapper/prior/__init__.py
@@ -1,0 +1,4 @@
+from .gaussian import GaussianPrior
+from .log_gaussian import LogGaussianPrior
+from .log_uniform import LogUniformPrior
+from .uniform import UniformPrior
