--- autofit/mapper/prior/prior.py
@@ -1,295 +1,0 @@
-import numpy as np
-
-from autofit import exc
-from autofit.messages.normal import NormalMessage, UniformNormalMessage, LogNormalMessage
-from autofit.messages.transform import log_10_transform
-from .abstract import Prior
-from .abstract import epsilon
-from .wrapped_instance import WrappedInstance
-
-
-class UniformPrior(Prior):
-    """A prior with a uniform distribution between a lower and upper limit"""
-
-    def __init__(
-            self,
-            lower_limit=0.0,
-            upper_limit=1.0,
-            id_=None,
-    ):
-        lower_limit = float(lower_limit)
-        upper_limit = float(upper_limit)
-
-        Message = UniformNormalMessage.shifted(
-            shift=lower_limit,
-            scale=upper_limit - lower_limit,
-        )
-        super().__init__(
-            WrappedInstance(
-                Message,
-                0, 1,
-                lower_limit=lower_limit,
-                upper_limit=upper_limit,
-                id_=id_,
-            ),
-            lower_limit=lower_limit,
-            upper_limit=upper_limit,
-            id_=id_,
-        )
-
-    def logpdf(self, x):
-        # TODO: handle x as a numpy array
-        if x == self.lower_limit:
-            x += epsilon
-        elif x == self.upper_limit:
-            x -= epsilon
-        return self.instance().logpdf(x)
-
-    def __str__(self):
-        """The line of text describing this prior for the model_mapper.info file"""
-        return f"UniformPrior, lower_limit = {self.lower_limit}, upper_limit = {self.upper_limit}"
-
-    def value_for(self, unit, ignore_prior_limits=False):
-        """
-
-        Parameters
-        ----------
-        unit: Float
-            A unit hypercube value between 0 and 1
-        Returns
-        -------
-        value: Float
-            A value for the attribute between the upper and lower limits
-        """
-        return round(super().value_for(unit, ignore_prior_limits=ignore_prior_limits), 14)
-
-    # noinspection PyUnusedLocal
-    @staticmethod
-    def log_prior_from_value(value):
-        """
-        Returns the log prior of a physical value, so the log likelihood of a model evaluation can be converted to a
-        posterior as log_prior + log_likelihood.
-
-        This is used by Emcee in the log likelihood function evaluation.
-
-        NOTE: For a UniformPrior this is always zero, provided the value is between the lower and upper limit. Given
-        this is check for when the instance is made (in the *instance_from_vector* function), we thus can simply return
-        zero in this function.
-        """
-        return 0.0
-
-
-class LogUniformPrior(Prior):
-    """A prior with a uniform distribution between a lower and upper limit"""
-
-    def __init__(
-            self,
-            lower_limit=1e-6,
-            upper_limit=1.0,
-            id_=None,
-    ):
-        if lower_limit <= 0.0:
-            raise exc.PriorException(
-                "The lower limit of a LogUniformPrior cannot be zero or negative."
-            )
-
-        lower_limit = float(lower_limit)
-        upper_limit = float(upper_limit)
-
-        Message = UniformNormalMessage.shifted(
-            shift=np.log10(lower_limit),
-            scale=np.log10(upper_limit / lower_limit),
-        ).transformed(
-            log_10_transform
-        )
-
-        super().__init__(
-            message=WrappedInstance(
-                Message,
-                0.0, 1.0,
-                id_=id_,
-                lower_limit=lower_limit,
-                upper_limit=upper_limit,
-            ),
-            lower_limit=lower_limit,
-            upper_limit=upper_limit,
-            id_=id_,
-        )
-
-    @classmethod
-    def with_limits(
-            cls,
-            lower_limit: float,
-            upper_limit: float
-    ):
-        return cls(
-            lower_limit=max(
-                0.000001,
-                lower_limit
-            ),
-            upper_limit=upper_limit,
-        )
-
-    __identifier_fields__ = ("lower_limit", "upper_limit")
-
-    @staticmethod
-    def log_prior_from_value(value):
-        """
-        Returns the log prior of a physical value, so the log likelihood of a model evaluation can be converted to a
-            posterior as log_prior + log_likelihood.
-
-        This is used by Emcee in the log likelihood function evaluation.
-
-        Parameters
-        ----------
-        value : float
-            The physical value of this prior's corresponding parameter in a `NonLinearSearch` sample."""
-        return 1.0 / value
-
-    def value_for(self, unit: float, ignore_prior_limits=False) -> float:
-        return super().value_for(unit, ignore_prior_limits=ignore_prior_limits)
-
-    def __str__(self):
-        """The line of text describing this prior for the model_mapper.info file"""
-        return f"LogUniformPrior, lower_limit = {self.lower_limit}, upper_limit = {self.upper_limit}"
-
-
-class GaussianPrior(Prior):
-    """A prior with a gaussian distribution"""
-
-    __identifier_fields__ = (
-        "lower_limit",
-        "upper_limit",
-        "mean",
-        "sigma"
-    )
-
-    def __init__(
-            self,
-            mean,
-            sigma,
-            lower_limit=float("-inf"),
-            upper_limit=float("inf"),
-
-    ):
-        super().__init__(
-            lower_limit=lower_limit,
-            upper_limit=upper_limit,
-            message=NormalMessage(
-                mean=mean,
-                sigma=sigma,
-                lower_limit=lower_limit,
-                upper_limit=upper_limit,
-            )
-        )
-
-    @classmethod
-    def with_limits(
-            cls,
-            lower_limit: float,
-            upper_limit: float
-    ) -> "GaussianPrior":
-        """
-        Create a new gaussian prior centred between two limits
-        with sigma distance between this limits.
-
-        Note that these limits are not strict so exceptions will not
-        be raised for values outside of the limits.
-
-        Parameters
-        ----------
-        lower_limit
-        upper_limit
-
-        Returns
-        -------
-        A new GaussianPrior
-        """
-        return cls(
-            mean=(lower_limit + upper_limit) / 2,
-            sigma=upper_limit - lower_limit,
-        )
-
-    def dict(self) -> dict:
-        """
-        A dictionary representation of this prior
-        """
-        prior_dict = super().dict()
-        return {**prior_dict, "mean": self.mean, "sigma": self.sigma}
-
-    def __str__(self):
-        """The line of text describing this prior for the model_mapper.info file"""
-        return f"GaussianPrior, mean = {self.mean}, sigma = {self.sigma}"
-
-
-class LogGaussianPrior(Prior):
-    """A prior with a log gaussian distribution"""
-
-    __identifier_fields__ = (
-        "lower_limit",
-        "upper_limit",
-        "mean",
-        "sigma"
-    )
-
-    def __init__(
-            self,
-            mean,
-            sigma,
-            lower_limit=0.0,
-            upper_limit=float("inf"),
-            id_=None,
-    ):
-        lower_limit = float(lower_limit)
-        upper_limit = float(upper_limit)
-
-        self.mean = mean
-        self.sigma = sigma
-
-        message = WrappedInstance(
-            LogNormalMessage,
-            mean=mean,
-            sigma=sigma,
-            id_=id_,
-            lower_limit=lower_limit,
-            upper_limit=upper_limit,
-        )
-        super().__init__(
-            message=message,
-            lower_limit=0.0,
-            upper_limit=float("inf"),
-            id_=None,
-        )
-
-    def _new_for_base_message(
-            self,
-            message
-    ):
-        """
-        Create a new instance of this wrapper but change the parameters used
-        to instantiate the underlying message. This is useful for retaining
-        the same transform stack after recreating the underlying message during
-        projection.
-        """
-        return LogGaussianPrior(
-            *message.parameters,
-            lower_limit=self.lower_limit,
-            upper_limit=self.upper_limit,
-            id_=self.instance().id
-        )
-
-    # @assert_within_limits
-    def value_for(self, unit, ignore_prior_limits=False):
-        """
-
-        Parameters
-        ----------
-        unit: Float
-            A unit hypercube value between 0 and 1
-
-        Returns
-        -------
-        value: Float
-            A value for the attribute biased to the gaussian distribution
-        """
-        return super().value_for(unit, ignore_prior_limits=ignore_prior_limits)
