+++ autofit/non_linear/grid/grid_search/__init__.py
@@ -5,7 +5,7 @@ from os import path
 from typing import List, Tuple, Union, Type, Optional, Dict
 
 from autofit import exc
-from autofit.mapper.prior import prior as p
+from autofit.mapper import prior as p
 from autofit.non_linear.abstract_search import NonLinearSearch
 from autofit.non_linear.parallel import Process
 from autofit.text.text_util import padding
